+++ src/input/handler/filters/KeyFilteringInputHandler.cpp
@@ -2,20 +2,6 @@
 #include "KeyFilteringInputHandler.h"
 #include "../../../lib/keycode/keycode.h"
 
-std::string linux_rawname_to_yaml_name(std::string linux_rawname) {
-    if (linux_rawname.rfind("LEFT", 0) == 0) {
-        linux_rawname = linux_rawname.substr(4, std::string::npos);
-    } else if (linux_rawname.rfind("RIGHT", 0) == 0) {
-        linux_rawname = linux_rawname.substr(5, std::string::npos);
-    }
-
-    if (linux_rawname == "CTRL") {
-        linux_rawname = "CONTROL";
-    }
-
-    return linux_rawname;
-}
-
 std::string KeyFilteringInputHandler::getFilterText() {
     std::string str;
 
@@ -29,7 +15,6 @@ std::string KeyFilteringInputHandler::getFilterText() {
 
     return str;
 }
-
 
 InputMode KeyFilteringInputHandler::getNextMode() {
     return InputMode::SELECTION;
@@ -45,7 +30,6 @@ std::function<bool(Hotkey *)> KeyFilteringInputHandler::getFilter() {
 
                     keyStr = linux_rawname_to_yaml_name(keyStr);
 
-                    // TODO: we need a translator from key name(from config file) to linux key name
                     if (hotkey->getKeyCodes()->find(keyStr) == hotkey->getKeyCodes()->end()) {
                         return false;
                     }
